pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = ""
authors = ["thachit <nguyencothach1989@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.9"
uvicorn = "^0.15.0"
fastapi = "^0.115.2"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
psycopg2 = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.9.2"}
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
